iocompat: adds eventMiddleware
This middleware forwards events from the browser so we can
keep our models here in sync, using bindBackend() as usual.

For requests that came over redis we do nothing as the model
already listens for the 'redis' events, we just pass it along
the chain.

Else, if the request came from the browser we emit a new pair
of events and keep churning.