iobackends: adds tmpId middleware
On the browser sometimes we need to set the '_id' of new objects so Backbone-relational
does not break (when we try to save new objects their relations end up pointing to 'undefined').

On that case we need to turn the 'update' request into a 'create' for other browsers to add the
newly created objects.

To signal this particular condition we add an '_tmpid' attribute to the model and catch that here.